+++ Epsilon.Abstractions/Component/KpiMatrixComponent/KpiMatrix.cs
@@ -1,4 +1,4 @@
-﻿namespace Epsilon.Abstractions.Component;
+﻿namespace Epsilon.Abstractions.Component.KpiMatrixComponent;
 
 public record KpiMatrix(
     IEnumerable<KpiMatrixOutcome> Outcomes,
