+++ Epsilon.Abstractions/Component/KpiMatrixComponent/KpiMatrixOutcome.cs
@@ -1,6 +1,6 @@
 using Epsilon.Canvas.Abstractions.Model;
 
-namespace Epsilon.Abstractions.Component;
+namespace Epsilon.Abstractions.Component.KpiMatrixComponent;
 public record KpiMatrixOutcome(
     int Id,
     string Title,
