+++ Epsilon.Host.WebApi/Program.cs
@@ -1,4 +1,5 @@
 using Epsilon.Abstractions.Component;
+using Epsilon.Abstractions.Component.KpiMatrixComponent;
 using Epsilon.Canvas;
 using Epsilon.Component;
